v0.3.0 - Real Deal ### Added - Added CI for project
### Changed
- Updated the project structure
