+++ README.md
@@ -1,3 +1,3 @@
 # Kraken
 
-Release it
+[![codecov](https://codecov.io/gh/Podium-Education/kraken/branch/main/graph/badge.svg?token=YAoPTsYWbB)](https://codecov.io/gh/Podium-Education/kraken)
